+++ zyte_spider_templates/spiders/ecommerce.py
@@ -1,9 +1,10 @@
+from __future__ import annotations
+
 from enum import Enum
-from typing import Any, Callable, Dict, Iterable, Optional, Union
+from typing import TYPE_CHECKING, Any, Callable, Dict, Iterable, Optional, Union, cast
 
 import scrapy
 from pydantic import BaseModel, ConfigDict, Field, model_validator
-from scrapy import Request
 from scrapy.crawler import Crawler
 from scrapy_poet import DummyResponse, DynamicDeps
 from scrapy_spider_metadata import Args
@@ -37,6 +38,10 @@ from ..params import (
     UrlsFileParam,
     UrlsParam,
 )
+
+if TYPE_CHECKING:
+    # typing.Self requires Python 3.11
+    from typing_extensions import Self
 
 
 @document_enum
@@ -198,7 +203,7 @@ class EcommerceSpider(Args[EcommerceSpiderParams], BaseSpider):
     }
 
     @classmethod
-    def from_crawler(cls, crawler: Crawler, *args, **kwargs) -> scrapy.Spider:
+    def from_crawler(cls, crawler: Crawler, *args, **kwargs) -> Self:
         spider = super(EcommerceSpider, cls).from_crawler(crawler, *args, **kwargs)
         parse_input_params(spider)
         spider._init_extract_from()
@@ -222,7 +227,7 @@ class EcommerceSpider(Args[EcommerceSpiderParams], BaseSpider):
             if self.args.crawl_strategy == EcommerceCrawlStrategy.direct_item
             else self.parse_navigation
         )
-        meta = {
+        meta: Dict[str, Any] = {
             "crawling_logs": {
                 "page_type": "product"
                 if self.args.crawl_strategy == EcommerceCrawlStrategy.direct_item
@@ -252,13 +257,13 @@ class EcommerceSpider(Args[EcommerceSpiderParams], BaseSpider):
                     f"Heuristics won't be used to crawl other pages which might have products."
                 )
 
-        return Request(
+        return scrapy.Request(
             url=url,
             callback=callback,
             meta=meta,
         )
 
-    def start_requests(self) -> Iterable[Request]:
+    def start_requests(self) -> Iterable[scrapy.Request]:
         if self.args.search_queries:
             for url in self.start_urls:
                 meta: Dict[str, Any] = {
@@ -266,7 +271,7 @@ class EcommerceSpider(Args[EcommerceSpiderParams], BaseSpider):
                 }
                 if self.args.extract_from == ExtractFrom.browserHtml:
                     meta["inject"] = [BrowserResponse]
-                yield Request(
+                yield scrapy.Request(
                     url=url,
                     callback=self.parse_search_request_template,
                     meta=meta,
@@ -280,7 +285,7 @@ class EcommerceSpider(Args[EcommerceSpiderParams], BaseSpider):
         response: DummyResponse,
         search_request_template: SearchRequestTemplate,
         dynamic: DynamicDeps,
-    ) -> Iterable[Request]:
+    ) -> Iterable[scrapy.Request]:
         probability = search_request_template.get_probability()
         if probability is not None and probability <= 0:
             return
@@ -294,7 +299,7 @@ class EcommerceSpider(Args[EcommerceSpiderParams], BaseSpider):
 
     def parse_navigation(
         self, response: DummyResponse, navigation: ProductNavigation
-    ) -> Iterable[Request]:
+    ) -> Iterable[scrapy.Request]:
         page_params = self._modify_page_params_for_heuristics(
             response.meta.get("page_params")
         )
@@ -310,7 +315,9 @@ class EcommerceSpider(Args[EcommerceSpiderParams], BaseSpider):
                     f"are no product links found in {navigation.url}"
                 )
             else:
-                yield self.get_nextpage_request(navigation.nextPage)
+                yield self.get_nextpage_request(
+                    cast(ProbabilityRequest, navigation.nextPage)
+                )
 
         if (
             self.args.crawl_strategy != EcommerceCrawlStrategy.pagination_only
@@ -336,6 +343,7 @@ class EcommerceSpider(Args[EcommerceSpiderParams], BaseSpider):
             else:
                 yield product
         else:
+            assert self.crawler.stats
             self.crawler.stats.inc_value("drop_item/product/low_probability")
             self.logger.info(
                 f"Ignoring item from {response.url} since its probability is "
@@ -343,9 +351,7 @@ class EcommerceSpider(Args[EcommerceSpiderParams], BaseSpider):
             )
 
     @staticmethod
-    def get_parse_navigation_request_priority(
-        request: Union[ProbabilityRequest, Request]
-    ) -> int:
+    def get_parse_navigation_request_priority(request: ProbabilityRequest) -> int:
         if (
             not hasattr(request, "metadata")
             or not request.metadata
@@ -356,7 +362,7 @@ class EcommerceSpider(Args[EcommerceSpiderParams], BaseSpider):
 
     def get_parse_navigation_request(
         self,
-        request: Union[ProbabilityRequest, Request],
+        request: ProbabilityRequest,
         callback: Optional[Callable] = None,
         page_params: Optional[Dict[str, Any]] = None,
         priority: Optional[int] = None,
@@ -379,7 +385,7 @@ class EcommerceSpider(Args[EcommerceSpiderParams], BaseSpider):
 
     def get_subcategory_request(
         self,
-        request: Union[ProbabilityRequest, Request],
+        request: ProbabilityRequest,
         callback: Optional[Callable] = None,
         page_params: Optional[Dict[str, Any]] = None,
         priority: Optional[int] = None,
@@ -401,7 +407,7 @@ class EcommerceSpider(Args[EcommerceSpiderParams], BaseSpider):
 
     def get_nextpage_request(
         self,
-        request: Union[ProbabilityRequest, Request],
+        request: ProbabilityRequest,
         callback: Optional[Callable] = None,
         page_params: Optional[Dict[str, Any]] = None,
     ):
@@ -420,7 +426,7 @@ class EcommerceSpider(Args[EcommerceSpiderParams], BaseSpider):
         priority = self.get_parse_product_request_priority(request)
 
         probability = request.get_probability()
-        meta = {
+        meta: Dict[str, Any] = {
             "crawling_logs": {
                 "name": request.name,
                 "probability": probability,